LPS-65683 move to new module
