+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-runtime-scripting-impl/src/main/java/com/liferay/portal/workflow/kaleo/runtime/scripting/internal/assignment/DRLScriptingTaskAssignmentSelector.java
@@ -12,22 +12,24 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.runtime.internal.action.executor;
+package com.liferay.portal.workflow.kaleo.runtime.scripting.internal.assignment;
 
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.resource.StringResourceRetriever;
 import com.liferay.portal.rules.engine.Fact;
 import com.liferay.portal.rules.engine.Query;
 import com.liferay.portal.rules.engine.RulesEngine;
 import com.liferay.portal.rules.engine.RulesResourceRetriever;
-import com.liferay.portal.workflow.kaleo.model.KaleoAction;
+import com.liferay.portal.workflow.kaleo.model.KaleoTaskAssignment;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
-import com.liferay.portal.workflow.kaleo.runtime.action.executor.ActionExecutor;
-import com.liferay.portal.workflow.kaleo.runtime.action.executor.ActionExecutorException;
+import com.liferay.portal.workflow.kaleo.runtime.assignment.BaseTaskAssignmentSelector;
+import com.liferay.portal.workflow.kaleo.runtime.assignment.TaskAssignmentSelector;
 import com.liferay.portal.workflow.kaleo.runtime.util.RulesContextBuilder;
 import com.liferay.portal.workflow.kaleo.runtime.util.WorkflowContextUtil;
 
 import java.io.Serializable;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
@@ -38,38 +40,31 @@ import org.osgi.service.component.annotations.Reference;
  * @author Michael C. Han
  */
 @Component(
-	immediate = true,
-	property = {"com.liferay.portal.workflow.kaleo.runtime.action.executor.language=drl"},
-	service = ActionExecutor.class
+	immediate = true, property = {"scripting.language=drl"},
+	service = TaskAssignmentSelector.class
 )
-public class DRLActionExecutor implements ActionExecutor {
+public class DRLScriptingTaskAssignmentSelector
+	extends BaseTaskAssignmentSelector {
 
 	@Override
-	public void execute(
-			KaleoAction kaleoAction, ExecutionContext executionContext)
-		throws ActionExecutorException {
-
-		try {
-			doExecute(kaleoAction, executionContext);
-		}
-		catch (Exception e) {
-			throw new ActionExecutorException(e);
-		}
-	}
-
-	protected void doExecute(
-			KaleoAction kaleoAction, ExecutionContext executionContext)
-		throws Exception {
+	public Collection<KaleoTaskAssignment> calculateTaskAssignments(
+			KaleoTaskAssignment kaleoTaskAssignment,
+			ExecutionContext executionContext)
+		throws PortalException {
 
 		List<Fact<?>> facts = _rulesContextBuilder.buildRulesContext(
 			executionContext);
 
+		String assigneeScript = kaleoTaskAssignment.getAssigneeScript();
+
 		RulesResourceRetriever rulesResourceRetriever =
 			new RulesResourceRetriever(
-				new StringResourceRetriever(kaleoAction.getScript()));
+				new StringResourceRetriever(assigneeScript));
+
+		Query query = Query.createStandardQuery();
 
 		Map<String, ?> results = _rulesEngine.execute(
-			rulesResourceRetriever, facts, Query.createStandardQuery());
+			rulesResourceRetriever, facts, query);
 
 		Map<String, Serializable> resultsWorkflowContext =
 			(Map<String, Serializable>)results.get(
@@ -77,6 +72,8 @@ public class DRLActionExecutor implements ActionExecutor {
 
 		WorkflowContextUtil.mergeWorkflowContexts(
 			executionContext, resultsWorkflowContext);
+
+		return getKaleoTaskAssignments(results);
 	}
 
 	@Reference
