+++ modules/apps/forms-and-workflow/portal-workflow/portal-workflow-kaleo-runtime-scripting-impl/src/main/java/com/liferay/portal/workflow/kaleo/runtime/scripting/internal/condition/DRLConditionEvaluator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.runtime.internal.notification.recipient.script;
+package com.liferay.portal.workflow.kaleo.runtime.scripting.internal.condition;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.resource.StringResourceRetriever;
@@ -20,10 +20,13 @@ import com.liferay.portal.rules.engine.Fact;
 import com.liferay.portal.rules.engine.Query;
 import com.liferay.portal.rules.engine.RulesEngine;
 import com.liferay.portal.rules.engine.RulesResourceRetriever;
-import com.liferay.portal.workflow.kaleo.model.KaleoNotificationRecipient;
+import com.liferay.portal.workflow.kaleo.model.KaleoCondition;
 import com.liferay.portal.workflow.kaleo.runtime.ExecutionContext;
-import com.liferay.portal.workflow.kaleo.runtime.notification.recipient.script.NotificationRecipientEvaluator;
+import com.liferay.portal.workflow.kaleo.runtime.condition.ConditionEvaluator;
 import com.liferay.portal.workflow.kaleo.runtime.util.RulesContextBuilder;
+import com.liferay.portal.workflow.kaleo.runtime.util.WorkflowContextUtil;
+
+import java.io.Serializable;
 
 import java.util.List;
 import java.util.Map;
@@ -36,27 +39,46 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true, property = {"scripting.language=drl"},
-	service = NotificationRecipientEvaluator.class
+	service = ConditionEvaluator.class
 )
-public class DRLNotificationRecipientEvaluator
-	implements NotificationRecipientEvaluator {
+public class DRLConditionEvaluator implements ConditionEvaluator {
 
 	@Override
-	public Map<String, ?> evaluate(
-			KaleoNotificationRecipient kaleoNotificationRecipient,
-			ExecutionContext executionContext)
+	public String evaluate(
+			KaleoCondition kaleoCondition, ExecutionContext executionContext)
 		throws PortalException {
+
+		List<Fact<?>> facts = _rulesContextBuilder.buildRulesContext(
+			executionContext);
 
 		RulesResourceRetriever rulesResourceRetriever =
 			new RulesResourceRetriever(
-				new StringResourceRetriever(
-					kaleoNotificationRecipient.getRecipientScript()));
-		List<Fact<?>> facts = _rulesContextBuilder.buildRulesContext(
-			executionContext);
+				new StringResourceRetriever(kaleoCondition.getScript()));
+
 		Query query = Query.createStandardQuery();
 
-		return _rulesEngine.execute(rulesResourceRetriever, facts, query);
+		Map<String, ?> results = _rulesEngine.execute(
+			rulesResourceRetriever, facts, query);
+
+		String returnValue = (String)results.get(_RETURN_VALUE);
+
+		Map<String, Serializable> resultsWorkflowContext =
+			(Map<String, Serializable>)results.get(
+				WorkflowContextUtil.WORKFLOW_CONTEXT_NAME);
+
+		WorkflowContextUtil.mergeWorkflowContexts(
+			executionContext, resultsWorkflowContext);
+
+		if (returnValue != null) {
+			return returnValue;
+		}
+
+		throw new IllegalStateException(
+			"Conditional did not return value for script " +
+				kaleoCondition.getScript());
 	}
+
+	private static final String _RETURN_VALUE = "returnValue";
 
 	@Reference
 	private RulesContextBuilder _rulesContextBuilder;
